requirements.txt
langchain
ipykernel
langchain-community
pypdf
langchain-text-splitters
sentence-transformers
chromadb
langchain-huggingface
langchain-chroma
langchain[google-genai]
langchain-google-genai
python-dotenv
langgraph
streamlit